Implement OPEV-0014
This commit introduces to go-vex the required changes to implement OPEV-0014:

It introduces a new Component type that is the base of the new expanded product
and subomponent types.

It also modifies the cannonical hash algorith to add the new fields in the component
data to the cstring that is used to compute the hash.

It also updates the tests where required.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>