Implement URL property from standard
The CSAF standard defines a URL property for remediation data.
This property was missing but is actively used by Red Hat CSAF
VEX to communicate the respective Red Hat Security Advisory
for the vulnerability. We at DT need this attribute to get this data for our internal database of security advisories.

Signed-off-by: Jonathan Heinz <Jonathan.Heinz@telekom.de>